% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/textComplexitySimilarity.R
\name{textComplexitySimilarity}
\alias{textComplexitySimilarity}
\title{textComplexitySimilarity}
\usage{
textComplexitySimilarity(text)
}
\arguments{
\item{text}{What do you want to compare? Input your text as a character string}
}
\value{
Numeric string giving the similarity of the inputed text to texts from Harold Bloom's The Western Canon
}
\description{
This function compares the complexity of an inputed text to that of texts from Harold Bloom's the Western Canon
}
\examples{
textSimilarityComplexity()
}
\author{
Sean Fischer and Ben Wach
}
\keyword{Canon,}
\keyword{Comparison}
\keyword{Complexity,}
\keyword{Western}

